Remix/bide started
Attributes are still needed. currently there are 4 buttons for moves and a 5th for bide. bide is done twice and increases damage but has no other function. attacks can be chosen and when 2 are chosen the attack is made. you then attack and the turn switches allowing the enemies to attack and a new move to be made. current the attack damage gets stacked but again no other functionality. attributes update coming soon